requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
airportsdata==20250909
annotated-types==0.7.0
anthropic==0.66.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
blake3==1.0.5
blinker==1.9.0
blis==1.2.1
boto3==1.40.32
botocore==1.40.32
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
cleantext==1.1.4
click==8.1.8
cloudpathlib==0.22.0
cloudpickle==3.1.1
codetiming==1.4.0
coloredlogs==15.0.1
colorful==0.5.7
compressed-tensors==0.9.2
confection==0.1.5
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
datasets==4.0.0
debugpy==1.8.16
decorator==5.2.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
eval_type_backport==0.2.2
executing==2.2.1
faiss-cpu==1.12.0
Farama-Notifications==0.0.4
fastapi==0.116.1
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
flash_attn==2.7.4.post1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.6.1
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
gym==0.26.2
gym-notices==0.1.0
gym-sokoban==0.0.6
gymnasium==1.2.0
h11==0.16.0
hf-xet==1.1.9
html2text==2025.4.15
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
interegular==0.3.3
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
linkify-it-py==2.0.3
llguidance==0.7.30
llvmlite==0.43.0
lm-format-enforcer==0.10.12
marisa-trie==1.3.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.13.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
memray==1.18.0
mistral_common==1.8.4
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.22.1
openai==1.107.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-exporter-prometheus==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
preshed==3.0.10
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycountry==24.6.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyecharts==2.0.8
pygame==2.6.1
Pygments==2.19.2
pyjnius==1.7.0
pylatexenc==2.10
pynvml==13.0.1
pyparsing==3.2.3
pyserini==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
-e git+https://github.com/ZihanWang314/RAGEN.git@5eb79b82cecc31db5662ba3c96b752405692410b#egg=ragen
rank-bm25==0.2.2
RapidFuzz==3.14.1
ray==2.51.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
sentencepiece==0.2.1
sentry-sdk==2.37.1
setuptools==78.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
swankit==0.2.4
swanlab==0.6.10
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensordict==0.6.2
textual==6.2.1
thefuzz==0.22.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.11.0
together==1.5.25
tokenizers==0.22.0
torch==2.6.0+cu124
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
triton==3.2.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
-e git+https://github.com/volcengine/verl.git@1e47e412a441bae8cd1152888f6822871f95dec5#egg=verl&subdirectory=../../verl
virtualenv==20.35.3
vllm==0.8.2
wandb==0.21.3
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
